story 17 frotend add user and read products
