+++ mockdb.json
@@ -1,3 +1,13 @@
 {
-    "demo":[{"name": "5", "email": "918238923"}]
-  }
+  "demo": [
+    { "id": 1, "name": "5", "email": "918238923" },
+    { "id": 2, "name": "messi", "email": "mansanchez@hotmail.com" },
+    { "id": 2, "name": "messi", "email": "mansanchez@hotmail.com" }
+  ],
+  "products":[
+    {"id":1, "sku":"0221212441123", "name":"Brandy Harvey", "description":"The best player", "price":100, "is_active":1, "created_at":"2011-02-11", "updated_at":"2011-02-11","stock":"1","photo_file_name":"http://placeimg.com/640/480/food"}
+
+
+    
+  ]
+}
