+++ package.json
@@ -13,6 +13,7 @@
     "@emotion/styled": "^11.6.0",
     "@mui/material": "^5.1.1",
     "axios": "^0.24.0",
+    "mui-image": "^0.1.3",
     "next": "latest",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
