+++ styles/Home.module.css
@@ -5,6 +5,7 @@
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  background-color:#44546A;
 }
 
 .main {
